Split HasComponent() into HasComponent and ProvidesComponent
This allows clients to distinguish between having the component (and
so we can write it) and providing the component (and therefore either
we have it or can generate it on the fly).

If A.HasComponent("face") and !A.HasComponent("boundary_face"), then
A.ProvidesComponent("boundary_face").

Also adds some assertions, which may throw on a non-const request for
boundary_face (which indicates we need to use the const version!)