Adds CV::ProvidesComponent() and uses it in vector math
Previously, if A & B are CVs,

A.Update(1, B, 1)

where A.HasComponent("boundary_face") and B.HasComponent("face") would
do nothing, because !B.HasComponent("boundary_face").  This should be
allowed, however, because B.ViewComponent("boundary_face") uses the
Vandelay map and all is fine.

This commit makes the logic:

A.HasComponent(comp) <-- a _writable_ view of comp is available (no change)
A.ProvidesComponent(comp) <-- a _readable_ view of comp is available (new)

and uses ProvidesComponent() where possible to let Update(), Dot(),
etc work through the Vandelay map.

Note also that it is probably implied that:
!(A.HasComponent("boundary_face") && A.HasComponent("face"))
(you can't have both) though I'm not sure whether this breaks anything
or not.

Several changes to code were also made to:

if (A.HasComponent("face")) {
  ...
} else if (A.HasComponent("boundary_face")) {
  ...
}

rather than two independent if conditionals.  Likely this doesn't
really matter, and both are now valid.