Implemended a workaroud for chemistry plus transport.
A proper tag is needed for tcc in the chemistry PK. The case of the transport PK
followed by the chemistry PK requires Tags::COPY, since the transport advances
from NEXT to COPY. The case of the chemistry PK followed by the transport PK
advances from NEXT to NEXT, hence it requires a different start tag.

Removed unused constructors.

The transport PK can extract a pointer to the chemistry PK using the
parameter "sibling N" where N is the order of the chemistry PK
inside an MPC.