Fixes weights in the support-operator scheme.
The original non-normalized corner weight was defined as the determinant
of the matrix of unit face normals. In 2D, this implies sine of the
corner angle which differs from the conventional scheme which uses area
of the corner triangle. Now, areas of corner triangles and volumes of
corner tets are used as non-normalized weigths.